+++ from_snowman/other-003_gcd_m32/ia32_elf/by_jeb.c
@@ -1,4 +1,4 @@
 
-int gcd(int param0, int param1) {
-    return !param1 ? param0: gcd(param1, (unsigned int)((unsigned long long)param0 | ((unsigned long long)(param0 >> 31) << 32)) % param1);
+int gcd(int param0, unsigned int param1) {
+    return !param1 ? param0: gcd((int)param1, (unsigned int)((unsigned long long)param0 | ((unsigned long long)(param0 >> 31) << 32)) % (int)param1);
 }
