+++ from_snowman/other-014_calls/x64_elf/by_jeb.c
@@ -1,6 +1,6 @@
 
-long g(int param0, int param1, int param2, int param3) {
-    return (unsigned long)(param3 + param2 + (param1 + param0));
+int g(int param0, int param1, int param2, int param3) {
+    return param3 + param2 + (param1 + param0);
 }
 
 long h() {
